ClientThread.submit argument order: jobid required
Introducing an optional jobid parameter solved a deadlock in client programs
that needed to synchronize on their own dict of jobids. But making this
optional before the *args and **kwargs didn't make sense. Now jobid is always
required. There may be a better way to handle this, but I haven't figured it
out yet.